[DDW-821] Renames add wallet page helpers
