+++ tests/Wallets/helpers/add-wallet.js
